Upgrades jsonwebtoken library to 4.2.2
There is a vulnerability in this module when the verification part is
expecting a token digitally signed with an asymetric key (RS/ES
family) of algorithms but instead the attacker send a token digitally
signed with a symmetric algorithm (HS* family).

The issue is because this library has the very same signature to
verify both type of tokens (parameter: `secretOrPublicKey`).

Ref:
https://github.com/auth0/node-jsonwebtoken/commit/1bb584bc382295eeb7ee8c4452a673a77a68b687