Fixes a defect when serializer may return dup objects in an array
When we was going through items array we did this in parallel, so we
may hit race-conditions when root element is not updated with one of
elements we iterate resulting in dup object in root node.

Additionally changed _.filter to _.any to find out if object is
already in the root.